Modificado formato general del codigo
