+++ HASH/documentacion.txt
@@ -1,0 +1,50 @@
+
+En esta carpeta se encuentran los modulos y archivos de todas las funciones y procedimientos
+para el manejo e implementacion de sistemas de mapas de java en C, las funciones estan 
+formuladas para uso general y el listado de abajo determina que accion realiza cada una:
+ 
+	-limpia:        Limpia el buffer.
+	-ascii:         Calcula el valor ascii de una cadena de caracteres.
+	-equals:        Compara el contenido de 2 mapas, si estos son iguales se regresa un
+			0, de  lo contrario, regresa un 1.
+	-keySet:        Regresa el nodo raiz de una lista encadenada simple que contiene las
+			llaves del mapa que se le ingreso a la funcion.
+	-put:           Ingresa un elemento nuevo a un mapa ingresado a la funcion y regresa
+			el contenido del elemento nuevo
+	-remove_entry:  Elimina un elemento de un mapa ingresado a la funcion  buscandolo por
+			su llave y regresa el contenido de este elemento.
+	-collection:    Regresa el nodo raiz de una lista encadenada simple que contiene los 
+			elementos (contenido y llave) del mapa ingresado a la funcion.
+	-containsValue: Esta funcion busca un contenido especifico dentro del mapa y regresa
+			un 0 si es que dicho contenido es encontrado o un 1 si no.
+	-entrySet:      Regresa el nodo raiz de una lista encadenada simple que contiene los
+			contenidos de los elementos del mapa que se le ingreso a la funcion.
+	-hashCode:      Regresa el numero de valor hash del mapa ingresado a la funcion.
+	-isEmpty:       Regresa un 0 si el mapa ingresado a la funcion no tiene elmentos 
+			dentro y un 1 si tiene elementos.
+	-putAll:        Convina todos los elementos de un mapa2 dentro de un mapa1 ingresado.
+	-clear:         Elimina todos los elementos de un mapa ingresado a la funcion.
+	-containsKey:   Esta funcion busca una llave especifica dentro del mapa y regresa
+			un 0 si es que esa llave es encontrada o un 1 si no.
+	-get:           Esta funcion regresa el contenido de un elemento buscado por su llave
+			dentro de un mapa especificado.
+	-size:          Esta funcion regresa el numero total de elementos que existen dentro 
+			de un mapa ingresado a la funcion.
+
+El sistema de mapas funciona mediante la implementacion de una tabla hash con resolucion de 
+colisiones via encadenamiento separado, las tablas y nodos de las funciones deben de ser 
+determinados por quien las llegue usar en el programa que deba usarlas.
+ 
+Los datos correspondientes a la funcionalidad, tipos de entradas y salidas de informacion de cada 
+funcion estan en sus documentaciones individuales.
+
+Las funciones se pueden encontrar en cada uno de los modulos correspondientes:
+	general: limpia,ascii,verificar	
+	mapas1:  equals,keySet,put,remove_entry,collection.
+	mapas2:  containsValue,entrySet,hashCode,isEmpty,putAll.
+	mapas2:  clear,containsKey,get,size.
+
+CREADO POR:
+	Luis Armando Vizcaino Magdaleno
+	José Antonio Sánchez Zaballa 
+
