+++ HASH/mapas1.c
@@ -1,188 +1,34 @@
 /**
- *================================================================================================
- *@file   mapas.c
- *
- *@brief  En este modulo se encuentran todas las funciones y procedimientos  para el manejo de
- *        sistemas de mapas de java en C, las funciones estan formuladas para uso general y el 
- *        listado de abajo determina que accion realiza cada una:
- *
- *        -limpia:        Limpia el buffer.
- *        -ascii:         Calcula el valor ascii de una cadena de caracteres.
- *        -equals:        Compara el contenido de 2 mapas, si estos son iguales se regresa un 
- *                        0, de  lo contrario, regresa un 1.
- *        -keySet:        Regresa el nodo raiz de una lista que contiene las llaves del mapa que
- *                        se le ingreso a la funcion.
- *        -put:           Ingresa un elemento nuevo a un mapa ingresado a la funcion y regresa
- *                        el contenido del elemento nuevo
- *        -remove_entry:  Elimina un elemento de un mapa ingresado a la funcion  buscandolo por
- *                        su llave y regresa el contenido de este elemento.
- *        -collection:    Regresa el nodo raiz de una lista encadenada simple que contiene los 
- *                        elementos del mapa ingresado a la funcion.
- *        -containsValue: Esta funcion busca un contenido especifico dentro del mapa y regresa
- *                        un 0 si es que dicho contenido es encontrado o un 1 si no.
- *        -entrySet:      Regresa el nodo raiz de una lista que contiene los contenidos del mapa
- *                        que se le ingreso a la funcion.
- *        -hashCode:      Regresa el numero de valor hash del mapa ingresado a la funcion.
- *        -isEmpty:       Regresa un 0 si el mapa ingresado a la funcion no tiene elmentos 
- *                        dentro y un 1 si tiene elementos.
- *        -putAll:        Convina todos los elementos de un mapa2 dentro de un mapa1 ingresado.
- *        -clear:         Elimina todos los elementos de un mapa ingresado a la funcion.
- *        -containsKey:   Esta funcion busca una llave especifica dentro del mapa y regresa
- *                        un 0 si es que esa llave es encontrada o un 1 si no.
- *        -get:           Esta funcion regresa el contenido de un elemento buscado por su llave
- *                        dentro de un mapa especificado.
- *        -size:          Esta funcion regresa el numero total de elementos que existen dentro 
- *                        de un mapa ingresado a la funcion.
+*================================================================================
+*@file   mapas1.c
+*
+*@brief  Este modulo contiene parte de las funciones correspondientes al 
+*        funcionamiento de mapas en C, aqui podran econtrar la documentacion 
+*        individual asi como el codigo de las siguientes funciones:
+*
+*        equals, keySet, put, remove_entry, collection.
+*
+*@author Luis Armando Vizcaino Magdaleno
+*/
 
- *        El sistema de mapas funciona mediante la implementacion de una tabla hash con resolucion
- *        de colisiones via encadenamiento separado, las tablas y nodos de las funciones deben de 
- *        ser determinados por quien las llegue usar en el programa que deba usarlas.
- *
- *        Los datos correspondientes a la funcionalidad, tipos de entradas y salidas de
- *        informacion de cada funcion estan en sus documentaciones individuales.
- *
- *@author Luis Armando Vizcaino Magdaleno
- *
- *
- *================================================================================================
- */
 
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#define T 240  //CONSTANTE DE TAMAÑO, TABLA HASH
 
-typedef struct mapa  //ESTRUCTURA PARA CONTENIDO DE MAPA
-{
-	char LLAVE[30];
-	char CONTENIDO[30];
-	struct mapa *sig;
-}MAPA;
+#include "bibliotecas.h"
 
 //FUNCIONES GENERALES DEL PROGRAMA
-void limpia(void);
-int ascii(char llave[]);
-void verificar(MAPA *nodo, MAPA *mapa[]);
+extern void limpia(void);
+extern int ascii(char llave[]);
+extern void verificar(MAPA *nodo, MAPA *mapa[]);
 
 //FUNCIONES DESARROLLADAS POR LUIS VIZCAINO
 int equals(MAPA *mapa1[], MAPA *mapa2[]); 
 void keySet(MAPA *mapa[], MAPA **raiz);
 char *put(MAPA *mapa[], char contenido[], char llave[]);  
 char *remove_entry(MAPA *mapa[], char llave[]);  
-void collection(MAPA *mapa[], MAPA **raiz);  
-
-// FUNCIONES DESARROLLADAS POR LAEL VILLAR
-int containsValue(MAPA *mapa[],char contenido[]);  
-void entrySet(MAPA *mapa[], MAPA **raiz);
-int hashCode(MAPA *mapa[]);  
-int isEmpty(MAPA *mapa[]);  
-void putAll(MAPA *mapa1[], MAPA *mapa2[]);  
-
-//FUNCIONES DESARROLLADAS POR JOSE
-void clear(MAPA *mapa[]);  
-int containsKey(MAPA *mapa[], char llave[]);  
-char *get(MAPA *mapa[], char llave[]);
-int size(MAPA *mapa[]);  
-
-//================================================================================================|||||||||||||||||||||||||||
-//FUNCIONES GENERALES DEL PROGRAMA
-
-/**
- *================================================================================
- *@brief     Este procedimiento limpia el buffer.
- *
- *@variables c: Limpiador de buffer.
-  *===============================================================================
- */
-void limpia(void)
-{
-	char c;
-  	while((c=getchar())!='\n');  //LIMPIA EL BUFFER
-  	return;
-}
-
-/**
- *================================================================================
- *@brief     Esta funcion tranforma una cadena de caracteres a la suma de los
- *           valores ascii de cada letra que contiene, luego, regresa dicho valor.
- *
- *@entradas  char llave[]; Arreglo a tranformar.
- *
- *@salidas   int valor: La suma de todos los valores ascii del arreglo.
- *
- *@variables count: Longitud de la cadena de caracteres.
- *           z:     Contador de FOR.
- *           valor: Suma de los valores ascii de la cadena.
- *
- *@author    Luis Armando Vizcaino Magdaleno
-  *===============================================================================
- */
-int ascii(char llave[])
-{
-	int count, z, valor=0;
-	count = strlen(llave);  //CALCULA LONGITUD DE CADENA
-	for(z=0; z<count; z++)
-	{
-  		valor = valor + llave[z];  //SUMA DE VALORES ASCII
- 	}
- 	return(valor);
-}
-
-/**
- *================================================================================
- *@brief     Este procedimiento sirve como apoyo para la funcion "put", esta
- *           verifica que la llave ingresada para valor de nuevo elemento de un
- *           mapa sea unica, y en el caso de que sea repetida, sustiuye el valor
- *           previo del elemento por el nuevo ingresado.
- *
- *@senradas  MAPA *nodo:   Elemento nuevo a ser comparado contra el resto de los 
- *                         elementos.
- *           MAPA *mapa[]: Mapa donde el nuevo elemento sera ingresado.   
- *
- *@variables *aux: Apuntador a la tabla hash para realizar el recorrido y la
- *                 comparacion.
- *
- *@author    Luis Armando Vizcaino Magdaleno
-  *===============================================================================
- */
-void verificar(MAPA *nodo, MAPA *mapa[])
-{
-	MAPA *aux;
-	int j;
-	for(j=0;j<T;j++) //RECORRIDO DE TABLA HASH
-	{
-		aux = mapa[j];
-		if(mapa[j] != NULL)
-		{
-			if(strcmp(aux->LLAVE,nodo->LLAVE) == 0)  //VERIFICACION DE LA LLAVE
-			{
-				printf("\n\tEsta clave ya fue usada... Sustitullendo valor\n\t");
-				printf("\tVALOR ANTERIOR: %s\n",aux->CONTENIDO);
-				strcpy(aux->CONTENIDO,nodo->CONTENIDO);
-				return;
-			}
-			if(aux->sig != NULL)
-			{
-				while(aux->sig != NULL)  //VERIFICACION DE NODOS LIGADOS
-				{
-					aux = aux->sig;
-					if(strcmp(aux->LLAVE,nodo->LLAVE) == 0)
-					{
-						printf("\n\tEsta clave ya fue usada... Sustitullendo valor\n\t");
-						printf("\tVALOR ANTERIOR: %s\n",aux->CONTENIDO);
-						strcpy(aux->CONTENIDO,nodo->CONTENIDO);
-						return;
-					}
-				}
-			}
-		}
-	}
-	return;
-}
+void collection(MAPA *mapa[], MAPA **raiz);
 
 //================================================================================================|||||||||||||||||||||||||||
 //FUNCIONES DESARROLLADAS POR LUIS VIZCAINO
-
 
 /**
  *================================================================================
@@ -587,48 +433,5 @@ void collection(MAPA *mapa[], MAPA **raiz)
 	}
 	return;
 }
-
-//================================================================================================|||||||||||||||||||||||||||
-// FUNCIONES DESARROLLADAS POR LAEL
-
-int containsValue(MAPA *mapa[],char contenido[])
-{
-	return;
-}  
-void entrySet(MAPA *mapa[], MAPA **raiz)
-{
-	return;
-}  
-int hashCode(MAPA *mapa[])
-{
-	return;
-}
-int isEmpty(MAPA *mapa[])
-{
-	return;
-}  
-void putAll(MAPA *mapa1[], MAPA *mapa2[])
-{
-	return;
-}
-//================================================================================================|||||||||||||||||||||||||||
-//FUNCIONES DESARROLLADAS POR JOSE
-
-void clear(MAPA *mapa[])
-{
-	return;
-}  
-int containsKey(MAPA *mapa[], char llave[])
-{
-	return;
-}  
-char *get(MAPA *mapa[], char llave[])
-{
-	return;
-}
-int size(MAPA *mapa[])
-{
-	return;
-}  
 
 //================================================================================================|||||||||||||||||||||||||||
